Bug Fix
Making multiple clicks/commands in a short amount of time causes the disks to be suspended in between multiple destination points, wanting to go to them all.
Fixed it to not allow clicks/commands to broadcast delegates and call functions if a move command is ongoing.
Might create a command queue system in the future for quality of life to allow this behavior without bugs.